backend/requirements.txt
fastapi
uvicorn
elasticsearch==8.18.1
pydantic
python-dotenv
